AIP-159 (feat):Implement multi-file type upload, encryption, and analysis

Closes AIP-159

See merge request dataalliance/products/itowe/product/rtl-rag-chatbot-api!15

+++ rtl_rag_chatbot_api/app.py
@@ -3,6 +3,7 @@ import logging
 import os
 import shutil
 import uuid
+from pathlib import Path
 
 import chromadb
 import uvicorn
@@ -10,11 +11,13 @@ from chromadb.config import Settings
 from fastapi import FastAPI, File, Form, HTTPException, UploadFile
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import JSONResponse
+from pydantic import BaseModel
 from starlette_exporter import PrometheusMiddleware, handle_metrics
 
 from configs.app_config import Config
 from rtl_rag_chatbot_api.chatbot.chatbot_creator import Chatbot
 from rtl_rag_chatbot_api.chatbot.gcs_handler import GCSHandler
+from rtl_rag_chatbot_api.chatbot.image_reader import analyze_images
 from rtl_rag_chatbot_api.common.embeddings import run_preprocessor
 from rtl_rag_chatbot_api.common.encryption_utils import encrypt_file
 from rtl_rag_chatbot_api.common.models import (
@@ -23,6 +26,14 @@ from rtl_rag_chatbot_api.common.models import (
     PreprocessRequest,
     Query,
 )
+
+
+class ImageAnalysisUpload(BaseModel):
+    file_id: str
+    original_filename: str
+    is_image: bool
+    analysis: str
+
 
 """
 Main FastAPI application for the RAG PDF API.
@@ -89,21 +100,21 @@ async def preprocess(request: PreprocessRequest):
     destination_file_path = "local_data/"
 
     try:
-        contain_multimedia = False
+        is_image = False
         if os.path.exists(chroma_db_path):
             logging.info(f"Embeddings are ready {file_id}")
-            # Load the metadata to check if it includes contain_multimedia
+            # Load the metadata to check if it includes is_image
             metadata_path = os.path.join(chroma_db_path, "metadata.json")
             if os.path.exists(metadata_path):
                 with open(metadata_path, "r") as f:
                     metadata = json.load(f)
-                contain_multimedia = metadata.get("contain_multimedia", False)
+                is_image = metadata.get("is_image", False)
             else:
-                contain_multimedia = False
+                is_image = False
             return {
                 "status": "Embeddings are ready",
                 "folder": file_id,
-                "contain_multimedia": contain_multimedia,
+                "is_image": is_image,
             }
 
         embeddings_prefix = f"{embeddings_folder}/{file_id}/"
@@ -113,18 +124,18 @@ async def preprocess(request: PreprocessRequest):
             logging.info(f"Downloading embeddings for {file_id} from GCS")
             gcs_handler.download_files_from_folder_by_id(file_id)
             logging.info(f"Embeddings downloaded for {file_id}")
-            # Load the metadata to get contain_multimedia info
+            # Load the metadata to get is_image info
             metadata_path = os.path.join(chroma_db_path, "metadata.json")
             if os.path.exists(metadata_path):
                 with open(metadata_path, "r") as f:
                     metadata = json.load(f)
-                contain_multimedia = metadata.get("contain_multimedia", False)
+                is_image = metadata.get("is_image", False)
             else:
-                contain_multimedia = False
+                is_image = False
             return {
                 "status": "Embeddings downloaded from GCS",
                 "folder": file_id,
-                "contain_multimedia": contain_multimedia,
+                "is_image": is_image,
             }
 
         logging.info(f"No embeddings found for {file_id}. Processing raw files.")
@@ -154,7 +165,7 @@ async def preprocess(request: PreprocessRequest):
         with open(local_metadata_path, "r") as f:
             metadata = json.load(f)
 
-        if contain_multimedia:
+        if is_image:
             pass
             # will implement this in seperate container and call api here
         else:
@@ -164,7 +175,7 @@ async def preprocess(request: PreprocessRequest):
                 file_id=file_id,
                 chroma_db_path=chroma_db_path,
                 chroma_db=db,
-                contain_multimedia=contain_multimedia,
+                is_image=is_image,
                 gcs_handler=gcs_handler,
             )
 
@@ -175,7 +186,7 @@ async def preprocess(request: PreprocessRequest):
         return {
             "status": "Files processed and embeddings created successfully",
             "folder": file_id,
-            "contain_multimedia": contain_multimedia,
+            "is_image": is_image,
         }
 
     except Exception as e:
@@ -189,12 +200,14 @@ async def chat(query: Query):
         file_id = query.file_id
 
         if file_id not in initialized_chatbots:
-            await initialize_chatbot(file_id, query.model_choice)
+            raise HTTPException(
+                status_code=404, detail="Chatbot not initialized for this file"
+            )
 
         chatbot = initialized_chatbots[file_id]
 
         response = chatbot.get_answer(query.text)
-
+        """
         # Check if the response includes a request to generate a chart or graph
         if (
             "generate chart" in query.text.lower()
@@ -202,7 +215,7 @@ async def chat(query: Query):
         ):
             chart_data = chatbot.generate_chart(query.text)
             return {"response": response, "chart_data": chart_data}
-
+        """
         return {"response": response}
     except Exception as e:
         print(f"Unexpected error in chat endpoint: {str(e)}")
@@ -234,23 +247,19 @@ async def cleanup_files():
 
 
 @app.post("/file/upload", response_model=FileUploadResponse)
-async def upload_file(
-    file: UploadFile = File(...), contain_multimedia: bool = Form(...)
-):
+async def upload_file(file: UploadFile = File(...), is_image: bool = Form(...)):
     try:
+        # Following logic, check if the file is already uploaded.
+        # If finds the file, it download the embeddings.
         original_filename = file.filename
         logging.info(f"Received file for upload: {original_filename}")
 
-        # Check if the file already exists in the "files-raw" bucket
-        logging.info(f"Checking if file {original_filename} already exists")
         existing_file_id = gcs_handler.find_existing_file(original_filename)
-        logging.info(f"Result of find_existing_file: {existing_file_id}")
 
         if existing_file_id:
             logging.info(
                 f"File {original_filename} already exists with ID: {existing_file_id}"
             )
-            # File already exists, download embeddings
             chroma_db_path = f"./chroma_db/{existing_file_id}"
             os.makedirs(chroma_db_path, exist_ok=True)
 
@@ -260,35 +269,42 @@ async def upload_file(
                     f"Embeddings downloaded for existing file: {existing_file_id}"
                 )
 
+                # Initialize the chatbot for the existing file
+                await initialize_chatbot(
+                    existing_file_id, "gpt_3_5_turbo"
+                )  # Use a default model
+                logging.info(
+                    f"Chatbot initialized for existing file: {existing_file_id}"
+                )
+
                 return FileUploadResponse(
-                    message="File already exists. Embeddings downloaded.",
+                    message="File already exists. Embeddings downloaded and chatbot initialized.",
                     file_id=existing_file_id,
                     original_filename=original_filename,
-                    contain_multimedia=contain_multimedia,
+                    is_image=is_image,
                 )
             except Exception as e:
-                logging.error(f"Error downloading embeddings: {str(e)}")
+                logging.error(
+                    f"Error downloading embeddings or initializing chatbot: {str(e)}"
+                )
                 raise HTTPException(
-                    status_code=500, detail="Error downloading embeddings"
+                    status_code=500, detail="Error processing existing file"
                 )
 
         logging.info(
             f"File {original_filename} is new. Proceeding with upload and processing."
         )
-        # If file doesn't exist, proceed with normal upload and processing
         file_id = str(uuid.uuid4())
+        file_extension = os.path.splitext(original_filename)[1]
         encrypted_filename = f"{original_filename}.encrypted"
+        temp_file_path = f"temp_{file_id}_{file_extension}"
 
-        # Create a temporary file to store the uploaded content
-        temp_file_path = f"temp_{file_id}_{original_filename}"
         with open(temp_file_path, "wb") as buffer:
             content = await file.read()
             buffer.write(content)
 
-        # Encrypt the file
         encrypted_file_path = encrypt_file(temp_file_path)
 
-        # Upload the encrypted file to GCS
         raw_files_folder = "files-raw"
         destination_blob_name = f"{raw_files_folder}/{file_id}/{encrypted_filename}"
         gcs_handler.upload_to_gcs(
@@ -296,7 +312,7 @@ async def upload_file(
             {
                 "file": (encrypted_file_path, destination_blob_name),
                 "metadata": (
-                    {"contain_multimedia": contain_multimedia},
+                    {"is_image": is_image},
                     f"{raw_files_folder}/{file_id}/metadata.json",
                 ),
             },
@@ -304,11 +320,9 @@ async def upload_file(
 
         logging.info(f"File uploaded to GCS: {destination_blob_name}")
 
-        # Clean up temporary files
         os.remove(temp_file_path)
         os.remove(encrypted_file_path)
 
-        # Prepare for preprocessing
         chroma_db_path = f"./chroma_db/{file_id}"
         os.makedirs(chroma_db_path, exist_ok=True)
         os.chmod(chroma_db_path, 0o755)
@@ -320,46 +334,58 @@ async def upload_file(
         destination_file_path = f"local_data/{file_id}/"
         os.makedirs(destination_file_path, exist_ok=True)
 
+        analysis_json_path = None
+        decrypted_file_path = None
         try:
-            # Download and decrypt the file
             decrypted_file_path = gcs_handler.download_and_decrypt_file(
                 file_id, destination_file_path
             )
             logging.info(f"File decrypted: {decrypted_file_path}")
 
-            # Preprocess the file
-            if contain_multimedia:
-                pass
-            else:
-                # Use the old code for text-only processing
-                run_preprocessor(
-                    configs=configs,
-                    text_data_folder_path=destination_file_path,
-                    file_id=file_id,
-                    chroma_db_path=chroma_db_path,
-                    chroma_db=db,
-                    contain_multimedia=contain_multimedia,
-                    gcs_handler=gcs_handler,
-                )
-                logging.info(f"Text processing completed for file: {file_id}")
+            # Check if the file is an image or PDF
+            if is_image:
+                logging.info("Processing image file: ..")
+                image_analysis_result = analyze_images(decrypted_file_path)
+                print(image_analysis_result)
+                if os.path.exists(decrypted_file_path):
+                    os.remove(decrypted_file_path)
 
-            # Initialize the chatbot after processing
-            await initialize_chatbot(file_id, "gpt_3_5_turbo")  # Use a default model
+                # Save the analysis result to a JSON file
+                analysis_json_path = os.path.join(
+                    destination_file_path, f"{file_id}_analysis.json"
+                )
+                with open(analysis_json_path, "w") as f:
+                    json.dump(image_analysis_result, f)
+
+            # Run preprocessor with the analysis JSON file
+            run_preprocessor(
+                configs=configs,
+                text_data_folder_path=destination_file_path,
+                file_id=file_id,
+                chroma_db_path=chroma_db_path,
+                chroma_db=db,
+                is_image=is_image,
+                gcs_handler=gcs_handler,
+            )
+            logging.info(f"Processing completed for file: {file_id}")
+
+            await initialize_chatbot(file_id, "gpt_3_5_turbo")
+            logging.info(f"Chatbot initialized for new file: {file_id}")
         except Exception as e:
             logging.error(f"Error during file processing: {str(e)}")
             raise
-
         finally:
-            # Clean up local files
-            if os.path.exists(decrypted_file_path):
+            if decrypted_file_path and os.path.exists(decrypted_file_path):
                 os.remove(decrypted_file_path)
+            if analysis_json_path and os.path.exists(analysis_json_path):
+                os.remove(analysis_json_path)
             shutil.rmtree(destination_file_path, ignore_errors=True)
 
         return FileUploadResponse(
-            message="File uploaded, encrypted, and preprocessed successfully",
+            message="File uploaded, encrypted, preprocessed, and chatbot initialized successfully",
             file_id=file_id,
             original_filename=original_filename,
-            contain_multimedia=contain_multimedia,
+            is_image=is_image,
         )
 
     except Exception as e:
@@ -401,7 +427,9 @@ async def get_neighbors(query: NeighborsQuery):
         file_id = query.file_id
 
         if file_id not in initialized_chatbots:
-            await initialize_chatbot(file_id)
+            raise HTTPException(
+                status_code=404, detail="Chatbot not initialized for this file"
+            )
 
         chatbot = initialized_chatbots[file_id]
 
@@ -418,9 +446,57 @@ async def get_neighbors(query: NeighborsQuery):
         raise HTTPException(status_code=500, detail=f"Unexpected error: {str(e)}")
 
 
+@app.post("/image/analyze")
+async def analyze_image_endpoint(file: UploadFile = File(...)):
+    try:
+        # Get the file extension from the original filename
+        file_extension = Path(file.filename).suffix
+
+        # Create a temporary file with the original extension
+        temp_file_path = f"temp_{uuid.uuid4()}{file_extension}"
+
+        with open(temp_file_path, "wb") as buffer:
+            content = await file.read()
+            buffer.write(content)
+
+        # Analyze the image
+        result = analyze_images(temp_file_path)
+
+        # Generate a unique filename for the result
+        result_filename = f"image_analysis_{uuid.uuid4()}.json"
+        result_file_path = os.path.join("processed_data", result_filename)
+
+        # Ensure the directory exists
+        os.makedirs(os.path.dirname(result_file_path), exist_ok=True)
+
+        # Save the result to a JSON file
+        with open(result_file_path, "w", encoding="utf-8") as f:
+            json.dump({"analysis": result}, f, ensure_ascii=False, indent=4)
+
+        # Clean up the temporary file
+        os.remove(temp_file_path)
+
+        return JSONResponse(
+            content={
+                "message": "Image analyzed successfully",
+                "result_file": result_filename,
+                "analysis": result,
+            }
+        )
+
+    except Exception as e:
+        logging.error(f"Error in image analysis: {str(e)}")
+        if os.path.exists(temp_file_path):
+            os.remove(temp_file_path)
+        raise HTTPException(
+            status_code=500, detail=f"An error occurred during image analysis: {str(e)}"
+        )
+
+
 def start():
     """
     Function to start the FastAPI application.
     Launched with `poetry run start` at root level
+    Streamlit : streamlit run streamlit_app.py
     """
     uvicorn.run("rtl_rag_chatbot_api.app:app", host="0.0.0.0", port=8080, reload=False)
